사용자 기능 및 후기 게시판 일부 기능 수정

+++ Uravel/WebContent/user/userinfo.jsp
@@ -1,43 +1,49 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8"
-    pageEncoding="UTF-8"%>
+	pageEncoding="UTF-8"%>
 
 <%@page import="java.util.*"%>
-<%@page import="com.dto.MemberDto"%>    
+<%@page import="com.dto.MemberDto"%>
 <%@page import="com.dto.UserThemeDto"%>
 <%@page import="com.dto.UserLocalDto"%>
 
-<% request.setCharacterEncoding("UTF-8"); %>    
-<% response.setContentType("text/html; charset=UTF-8"); %>
-    
+<%
+request.setCharacterEncoding("UTF-8");
+%>
+<%
+response.setContentType("text/html; charset=UTF-8");
+%>
+
 <!DOCTYPE html>
 <html>
 <head>
 <meta charset="UTF-8">
 <title>Insert title here</title>
-<link rel="stylesheet" href="../css/style.css">
-<link rel="stylesheet" href="../css/header.css">
-<link rel="stylesheet" href="../css/mypage.css">
+<link rel="stylesheet"
+	href="${pageContext.request.contextPath }/css/style.css">
+<link rel="stylesheet"
+	href="${pageContext.request.contextPath }/css/header.css">
 
 <script type="text/javascript">
 	function updateuser(userno){
 		location.href="logincontroller.jsp?command=updateform&userno="+userno;
 	}
 	function deleteuser(userno){
-		location.href="logincontroller.jsp?command=deleteuser&userno="+userno;
+		var returnValue = confirm("정말로 탈퇴하시겠습니까?");
+		if(returnValue == true){
+			location.href="logincontroller.jsp?command=deleteuser&userno="+userno;
+		}
 	}
 </script>
 </head>
+<%
+MemberDto userinfo = (MemberDto) request.getAttribute("dto");
+List<UserThemeDto> resultThemeList = (List<UserThemeDto>) request.getAttribute("resultUserThemeList");
+List<UserLocalDto> resultLocalList = (List<UserLocalDto>) request.getAttribute("resultUserLocalList");
+%>
 <body>
 	<%@ include file="../common/header.jsp"%>
 	<main>
-
-		<%@ include file="./user_side.jsp"%>
-<%
-	MemberDto userinfo = (MemberDto)request.getAttribute("dto");
-	List<UserThemeDto> resultThemeList = (List<UserThemeDto>)request.getAttribute("resultUserThemeList"); 
-	List<UserLocalDto> resultLocalList = (List<UserLocalDto>)request.getAttribute("resultUserLocalList");
-%>
-		
+		<%@ include file="./usertitle.jsp"%>
 		<div class="admin-main">
 			<div class="travel-info">
 				<h1>내 정보 관리</h1>
@@ -47,31 +53,47 @@
 					<table>
 						<colgroup>
 							<col width="30%">
-							<col width="100px">
+							<col width="800px">
 						</colgroup>
 						<tr>
 							<th>아이디</th>
 							<td><%=userinfo.getUserid()%></td>
 						</tr>
 						<tr>
-							<th>비밀번호</th>
-							<td><%=userinfo.getUserpw()%></td>
-						</tr>
-						<tr>
 							<th>이름</th>
 							<td><%=userinfo.getUsername()%></td>
 						</tr>
+						<%
+						String birth = userInfo.getBirth();
+						String birth_var;
+
+						if (birth == null || birth.equals("null")) {
+							birth_var = "";
+						} else {
+							birth_var = birth;
+						}
+						%>
 						<tr>
 							<th>생년월일</th>
-							<td><%=userinfo.getBirth()%></td>
+							<td><%=birth_var%></td>
 						</tr>
 						<tr>
 							<th>이메일</th>
 							<td><%=userinfo.getEmail()%></td>
 						</tr>
+						<%
+						String phone = userInfo.getPhone();
+						String phone_var;
+
+						if (phone == null || phone.equals("null")) {
+							phone_var = "";
+						} else {
+							phone_var = phone;
+						}
+						%>
 						<tr>
 							<th>핸드폰</th>
-							<td><%=userinfo.getPhone()%></td>
+							<td><%=phone_var%></td>
 						</tr>
 						<tr>
 							<th>등급</th>
@@ -84,22 +106,32 @@
 						<tr>
 							<th>선호 테마</th>
 							<td>
-							<% for(int i=0; i<resultThemeList.size(); i++){ %>
-								<%if(i != 0){ %>,<%} %>
-								<%=resultThemeList.get(i).getThemename() %>	 	
-							<% } %>
+								<%
+								for (int i = 0; i < resultThemeList.size(); i++) {
+								%> <%
+ if (i != 0) {
+ %>,<%
+ }
+ %> <%=resultThemeList.get(i).getThemename()%> <%
+ }
+ %>
 							</td>
 						</tr>
 						<tr>
 							<th>선호 지역</th>
 							<td>
-							<% for(int i=0; i<resultLocalList.size(); i++){ %>
-								 <%if(i != 0){ %>,<%} %>
-								 <%=resultLocalList.get(i).getLocalname() %>	 	
-							<% } %>
-							
+								<%
+								for (int i = 0; i < resultLocalList.size(); i++) {
+								%> <%
+ if (i != 0) {
+ %>,<%
+ }
+ %> <%=resultLocalList.get(i).getLocalname()%> <%
+ }
+ %>
+
 							</td>
-							
+
 						</tr>
 						<tr>
 							<td colspan="2" align="right">
